django-back/requirements.txt
asgiref==3.5.2
boto3==1.26.24
botocore==1.29.24
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
Pillow==9.3.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
